+++ src/usr/install.rs
@@ -70,13 +70,14 @@ pub fn copy_files(verbose: bool) {
     create_dev("/dev/net/gw", "net-gw", verbose);
     create_dev("/dev/net/ip", "net-ip", verbose);
     create_dev("/dev/net/mac", "net-mac", verbose);
-    create_dev("/dev/net/usage", "net-usage", verbose);
+    create_dev("/dev/net/stat", "net-stat", verbose);
     create_dev("/dev/null", "null", verbose);
     create_dev("/dev/pipe", "pipe", verbose);
     create_dev("/dev/proc/id", "proc-id", verbose);
     create_dev("/dev/proc/dir", "proc-dir", verbose);
     create_dev("/dev/proc/env", "proc-env", verbose);
     create_dev("/dev/proc/user", "proc-user", verbose);
+    create_dev("/dev/proc/stat", "proc-stat", verbose);
     create_dev("/dev/random", "random", verbose);
     create_dev("/dev/snd/buffer", "snd-buffer", verbose);
     create_dev("/dev/speaker", "speaker", verbose);